Speech Recognition and minor issues fixed
Speech recognition now works after installing the PyAudio module properly.
Trying to exit with command "exit" was making an infinite bucle caused by the name of the function that shutdowns the system. It was renamed and now it works as it should.
Fixed issue of getting the same line while running the program. Now it asks and responds as it should depending on the instruction given by the user.