Talk function and error handling improved
The talk(text) function was improved. Now it is able to reproduce the message and print it on the console, or reproduce a message and don't print it, or reproduce a message and specify what to print on the console. Now the funtion is called with talk(text, console=True, consoleText=""), the console parameter is set to True by default so it will always print to the console unless it's set to False, and the consoleText will be printed only if console=False and consoleText is not an empty string.
With this the print() calls were reduced within the code, so is more practical to use.

The who is command was rising an error that was killing the program if the wikipedia page wasn't found. Now it handles the error by printing a message on the console notifying about a result not found, so the program doesn't need to be restarted to work.